+++ app/components/Header/GetStarted/index.js
@@ -1,0 +1,1 @@
+export { default as GetStartedHeader } from "./GetStarted";
