+++ app/components/Header/Standard/index.js
@@ -1,0 +1,1 @@
+export { default as StandardHeader } from "./Standard";
