+++ app/components/SideBar/Logo/Logo.jsx
@@ -1,7 +1,7 @@
 import { FormattedMessage as T } from "react-intl";
 import { Tooltip } from "shared";
 import { TESTNET, MAINNET } from "constants";
-import style from "./Logo.module.css";
+import styles from "./Logo.module.css";
 
 const Logo = React.memo(
   ({
@@ -12,7 +12,7 @@ const Logo = React.memo(
     isWatchingOnly,
     getRunningIndicator
   }) => (
-    <div className={expandSideBar ? style.logo : style.reducedLogo}>
+    <div className={expandSideBar ? styles.logo : styles.reducedLogo}>
       {isWatchingOnly && (
         <Tooltip
           text={
@@ -21,14 +21,14 @@ const Logo = React.memo(
               m="This is a watch-only wallet with limited functionality."
             />
           }>
-          <div className={style.watchOnlyIcon} />
+          <div className={styles.watchOnlyIcon} />
         </Tooltip>
       )}
       <button
         aria-label="Logo"
         onClick={!expandSideBar ? onExpandSideBar : null}
         className={
-          !expandSideBar ? style.hamburger : isTestNet ? TESTNET : MAINNET
+          !expandSideBar ? styles.hamburger : isTestNet ? styles[TESTNET] : styles[MAINNET]
         }
       />
       {getRunningIndicator && (
@@ -39,13 +39,13 @@ const Logo = React.memo(
               m="One or more of the following decrediton's features running in the background: Privacy Mixer, Ticket Auto Buyer, Purchase Ticket Attempt"
             />
           }>
-          <div className={style.mixerOn} />
+          <div className={styles.mixerOn} />
         </Tooltip>
       )}
       {expandSideBar && (
         <button
           aria-label="Reduce sidebar"
-          className={style.reducedArrow}
+          className={styles.reducedArrow}
           onClick={onReduceSideBar}
         />
       )}
