+++ app/components/layout/TitleHeader/TitleHeader.module.css
@@ -1,0 +1,91 @@
+.container {
+  width: 750px;
+  position: relative;
+  justify-content: space-between;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.title {
+  margin-bottom: 3px;
+  color: var(--stroke-color-hovered);
+  font-size: 27px;
+  line-height: 30px;
+}
+
+.icon {
+  margin-right: 10px;
+  margin-top: 1px;
+}
+
+.icon > div {
+  height: 30px;
+  width: 30px;
+  background-size: 30px;
+  background-repeat: no-repeat;
+}
+
+.accounts {
+  background-image: var(--header-accounts);
+}
+.security {
+  background-image: var(--header-security);
+}
+.settings {
+  background-image: var(--menu-settings);
+}
+.tickets {
+  background-image: var(--header-tickets);
+}
+.transactions {
+  background-image: var(--header-transactions);
+}
+.governance {
+  background-image: var(--header-governance);
+}
+.trezor {
+  background-image: var(--header-trezor);
+}
+.ln {
+  background-image: var(--header-lightning);
+}
+.tx-detail-icon-ticket {
+  background-image: var(--ticket-live-icon);
+}
+.tx-detail-icon-vote {
+  background-image: var(--ticket-voted-icon);
+}
+.tx-detail-icon-revocation {
+  background-image: var(--ticket-revoked-icon);
+}
+.tx-detail-icon-ticketfee {
+  background-image: var(--ticketfee-transaction-icon);
+}
+.tx-detail-icon-mixed {
+  background-image: var(--mixed-transaction-icon);
+}
+.tx-detail-icon-self {
+  background-image: var(--self-transaction-icon);
+}
+.tx-detail-icon-out {
+  background-image: var(--minus-big-icon);
+}
+.tx-detail-icon-in {
+  background-image: var(--plus-big-icon);
+}
+
+.button {
+  margin-left: auto;
+}
+
+@media screen and (max-width: 1106px) {
+  .container {
+    width: 649px;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  .container {
+    width: 345px;
+  }
+}
