+++ app/components/layout/TitleHeader/index.js
@@ -1,0 +1,1 @@
+export { default as TitleHeader } from "./TitleHeader";
