+++ app/components/layout/index.js
@@ -1,7 +1,7 @@
 export { default as TabbedPage } from "./TabbedPage";
 export { TabbedPageTab } from "./TabbedPage";
 export { default as StandaloneHeader } from "./StandaloneHeader";
-export { default as TitleHeader } from "./TitleHeader";
+export * from "./TitleHeader";
 export { default as DescriptionHeader } from "./DescriptionHeader";
 export { default as StandalonePageBody } from "./StandalonePageBody";
 export { default as StandalonePage } from "./StandalonePage";
