+++ app/components/views/TicketsPage/TicketsPage.module.css
@@ -1,0 +1,12 @@
+.smallBalance {
+  display: inline;
+  font-size: 11px;
+  font-weight: 600;
+  line-height: 14px;
+  padding: 3px;
+  color: var(--input-color-default);
+  background-color: var(--blue-highlight-background);
+  margin-left: 5px;
+  margin-right: 5px;
+  border-radius: 3px;
+}
