--- app/style/Header.less
@@ -1,139 +1,0 @@
-@import (reference) "./main.less";
-
-.header-small-balance {
-  display: inline;
-  font-size: 11px;
-  font-weight: 600;
-  line-height: 14px;
-  padding: 3px;
-  color: var(--input-color-default);
-  background-color: var(--blue-highlight-background);
-  margin-left: 5px;
-  margin-right: 5px;
-  border-radius: 3px;
-}
-
-.header-get-started {
-  padding-right: 42px;
-  padding-left: 80px;
-  background-color: var(--disabled-background-color);
-}
-
-.header-top-get-started {
-  height: 34px;
-  padding-top: 5px;
-  text-align: center;
-}
-
-.header-title-overview-get-started {
-  height: 54px;
-  padding-top: 13px;
-  color: var(--background-back-color);
-  font-size: 27px;
-}
-
-.header-meta-overview-get-started {
-  height: 54px;
-  padding-top: 5px;
-  font-size: 18px;
-  color: var(--stroke-color-hovered);
-  font-family: var(--font-family-monospace);
-}
-
-.header-top {
-  text-align: center;
-}
-
-.header-title-overview {
-  padding-top: 14px;
-  color: var(--stroke-color-hovered);
-  font-size: 27px;
-}
-
-.header-meta-overview {
-  height: 54px;
-  padding-top: 5px;
-  font-size: 53px;
-}
-
-.title-header-icon > div {
-  height: 30px;
-  width: 30px;
-  background-size: 30px;
-  background-repeat: no-repeat;
-
-  &.accounts {
-    background-image: var(--header-accounts);
-  }
-  &.security {
-    background-image: var(--header-security);
-  }
-  &.settings {
-    background-image: var(--menu-settings);
-  }
-  &.tickets {
-    background-image: var(--header-tickets);
-  }
-  &.transactions {
-    background-image: var(--header-transactions);
-  }
-  &.governance {
-    background-image: var(--header-governance);
-  }
-  &.trezor {
-    background-image: var(--header-trezor);
-  }
-  &.ln {
-    background-image: var(--header-lightning);
-  }
-
-  &.tx-detail-icon-ticket {
-    background-image: var(--ticket-live-icon);
-  }
-  &.tx-detail-icon-vote {
-    background-image: var(--ticket-voted-icon);
-  }
-  &.tx-detail-icon-revocation {
-    background-image: var(--ticket-revoked-icon);
-  }
-  &.tx-detail-icon-ticketfee {
-    background-image: var(--ticketfee-transaction-icon);
-  }
-  &.tx-detail-icon-mixed {
-    background-image: var(--mixed-transaction-icon);
-  }
-  &.tx-detail-icon-self {
-    background-image: var(--self-transaction-icon);
-  }
-
-  &.tx-detail-icon-out {
-    background-image: var(--minus-big-icon);
-  }
-
-  &.tx-detail-icon-in {
-    background-image: var(--plus-big-icon);
-  }
-}
-
-.mainnet {
-  border: none;
-  outline: none;
-  float: left;
-  height: 20px;
-  width: 150px;
-  background-position: left center;
-  background-repeat: no-repeat;
-  background-image: var(--mainnet-logo);
-  background-color: transparent;
-}
-
-.testnet {
-  border: none;
-  outline: none;
-  height: 20px;
-  width: 150px;
-  background-position: left center;
-  background-repeat: no-repeat;
-  background-image: var(--testnet-logo);
-  background-color: transparent;
-}
